rearquitects flask setup
